changed process vulnerability to process occurance, and moved filename creation to saveVulnToBucket
