+++ devsecops/artifact-registry-vulnerability-scanning/cloud-function/index.js
@@ -1,7 +1,8 @@
 const { ContainerAnalysisClient } = require('@google-cloud/containeranalysis');
 const functions = require('@google-cloud/functions-framework');
-const { processVulnerability } = require('./src/processVunerability.js');
+// const { processVulnerability } = require('./src/processVunerability.js');
 const { writeVulnToBucket } = require('./src/writeVulnToBucket.js');
+const { processOccurrence } = require('./src/processOccurance.js');
 
 const containerAnalysisClient = new ContainerAnalysisClient();
 
@@ -26,17 +27,22 @@ exports.imageVulnPubSubHandler = async (event, context) => {
       name: occurrenceName,
     });
 
-    // We're only interested in vunerability types - filter for these and process (which includes saving to bucket)
-    //  Using others for logging - can remove in future.
+    // We're only interested in the 'vunerability' type - filter for these then process and save to bucket
     switch (occurrence.kind) {
       case 'VULNERABILITY':
-        // console.log('VULNERABILITY occurrence');
+        console.log('VULNERABILITY occurrence');
         try {
-          const { bucketName, vulnOccurrence, occurenceFilename } = await processVulnerability(occurrence, bucketName);
+          const { vulnId, affectedPackage, affectedVersion, resourceName } = await processOccurrence(occurrence);
+          // const { vulnOccurrence, occurenceFilename } = await processVulnerability(occurrence);
           await writeVulnToBucket(
             bucketName,
-            JSON.stringify(vulnOccurrence),
-            occurenceFilename)
+            // JSON.stringify(vulnOccurrence),
+            JSON.stringify(occurrence),
+            vulnId,
+            affectedPackage,
+            affectedVersion,
+            resourceName)
+            // occurenceFilename)
           } catch (error) {
             console.log('Error processing vunerability:', error)
           }
