+++ devsecops/artifact-registry-vulnerability-scanning/cloud-function/src/writeVulnToBucket.js
@@ -3,13 +3,16 @@ const storage = new Storage();
 
 const writeVulnToBucket = async (
   bucketName,
-  vulnText,
-  destinationObjectName,
+  vulnText, //aka occurance
+  vulnId,
+  affectedPackage,
+  affectedVersion,
+  resourceName,
 ) => {
   try {
+    const destinationObjectName = `vulnerabilities/${vulnId}-${affectedPackage}-${affectedVersion}-${resourceName}`;
     const bucket = storage.bucket(bucketName);
     const file = bucket.file(destinationObjectName);
-
     await file.save(vulnText);
     console.log(
       `Saved ${destinationObjectName} successfully to ${bucketName} bucket`,
